update dir for buildoutput and sdimg

+++ script/imgscr.sh
@@ -3,6 +3,7 @@ CONFIG_BOARD=$MYBOARD
 echo "making $CONFIG_BOARD sdimg fat:${1}M rootfs:${2}M"
 BOOTPARTSIZE=$1
 FSPARTSIZE=$2
+cd $NPWD
 mkdir -p images
 cd images
 mkdir -p bootdir
@@ -18,18 +19,17 @@ if [ $FSPARTSIZE -lt 2 ];then
         exit 1
 fi
 
-if [ "${CONFIG_BOARD}_x" == "bbb_x" ];then
-        TABLESIZE=2
-else
-        TABLESIZE=1
-fi
+TABLESIZE=1
 FATSTART=$[ ${TABLESIZE} * 1024 * 1024 / 512 ]
+EXTSTART=$[ ( ${TABLESIZE} + ${BOOTPARTSIZE} )* 1024 * 1024 / 512 ]
 TOTALSIZE=$[ ($BOOTPARTSIZE + $FSPARTSIZE + $TABLESIZE)*1024 + 10 ]
-dd if=/dev/zero of=${SDIMG} bs=1k count=$TOTALSIZE > /dev/null 2>&1
-dd if=$SPLPATH of=${SDIMG} conv=notrunc seek=1 bs=${SPLOFFSET}k > /dev/null 2>&1
-dd if=$BOOTPATH of=${SDIMG} conv=notrunc seek=1 bs=${BOOTOFFSET}k > /dev/null 2>&1
 
-echo "tablesize : $TABLESIZE fat : $FATSTART ext4 : $EXTSTART total : $TOTALSIZE/1024"
+dd if=/dev/zero of=${SDIMG} bs=1k count=$TOTALSIZE > /dev/null 2>&1
+dd if=$SPLPATH of=${SDIMG} conv=notrunc seek=${SPLOFFSET} bs=1k > /dev/null 2>&1
+dd if=$BOOTPATH of=${SDIMG} conv=notrunc seek=${BOOTOFFSET} bs=1k > /dev/null 2>&1
+sync
+
+# echo "tablesize: $TABLESIZE M fat: $FATSTART ext4: $EXTSTART total : $TOTALSIZE/1024"
 #create img partition table
 cat << FDISK_EOF | sudo fdisk ${SDIMG} > /dev/null 2>&1
 n
@@ -40,7 +40,7 @@ ${FATSTART}
 n
 p
 2
-
+${EXTSTART}
 +${FSPARTSIZE}M
 t
 1
@@ -56,10 +56,10 @@ if [ $? -ne 0 ]; then
         exit 1
 fi
 sync
-sleep 2
+sleep 1
 partprobe -s ${SDIMG}  > /dev/null 2>&1
 
-sudo kpartx -av ${SDIMG}
+sudo kpartx -av ${SDIMG}  > /dev/null 2>&1
 BOOTPART=` ls /dev/mapper | grep "p1$" `
 ROOTPART=` ls /dev/mapper | grep "p2$" `
 if [ "$BOOTPART" == "" -o "$ROOTPART" == "" ]; then
@@ -77,6 +77,10 @@ if ! sudo mount /dev/mapper/$BOOTPART bootdir; then
         sudo kpartx -dv ${SDIMG}
         exit 1
 fi
+# if [ "${CONFIG_BOARD}_x" == "bbb_x" ];then
+#         sudo cp -rf $SPLPATH $NPWD/images/bootdir/
+#         sudo cp -rf $BOOTPATH $NPWD/images/bootdir/
+# fi
 sudo cp -rf $IMGPATH $NPWD/images/bootdir/
 sudo cp -rf $DTBPATH $NPWD/images/bootdir/
 sudo cp -rf $ENVPATH $NPWD/images/bootdir/
@@ -91,8 +95,13 @@ if ! sudo mount -t ext4 /dev/mapper/$ROOTPART linuxdir; then
         echo "ERROR mounting rootfs linux partitions..."
         exit 1
 fi
-mkdir -p $NPWD/tmp
-mv  $NPWD/rootfs/root/*  $NPWD/tmp
+
+HAVEROOTFILE=`ls -al $NPWD/rootfs/root | awk '{if (NR > 3) {print $NF } }'`
+if [ "$HAVEROOTFILE" != "" ]; then
+        mkdir -p $NPWD/tmp
+        mv  $NPWD/rootfs/root/*  $NPWD/tmp
+fi
+
 sudo rsync -r -t -p -o -g -x --delete -l -H -D --numeric-ids -s --stats $NPWD/rootfs/ linuxdir > /dev/null 2>&1
 if [ $? -ne 0 ]; then
         echo "ERROR copying rootfs linux partition, maybe not enough size for rootfs"
@@ -104,11 +113,14 @@ sync
 if ! sudo umount linuxdir; then
         echo "ERROR unmounting rootfs linux partitions."
 fi
-sudo kpartx -dv ${SDIMG}
-mv $NPWD/tmp/* $NPWD/rootfs/root/ 
-rm -rf $NPWD/tmp
+sudo kpartx -dv ${SDIMG}  > /dev/null 2>&1
 
-cat << FDISK_EOF | sudo fdisk ${SDIMG} #> /dev/null 2>&1
+if [ "$HAVEROOTFILE" != "" ]; then
+        mv $NPWD/tmp/* $NPWD/rootfs/root/ 
+        rm -rf $NPWD/tmp
+fi
+
+cat << FDISK_EOF | sudo fdisk ${SDIMG} | grep "sd.img" #> /dev/null 2>&1
 p
 q
 FDISK_EOF
